Move Settings to its own package
